+++ hcse/__init__.py
@@ -1,9 +1,6 @@
-from importlib import import_module
+"""Holland Consciousness Scaling Engine package."""
 
-# Re-export submodule from nested package for backward compatibility
-_submodule = import_module('hcse.hcse')
+from .core import HCSEMixin
+from .pipeline import HfTrainerWithHCSE
 
-for attr in getattr(_submodule, '__all__', []):
-    globals()[attr] = getattr(_submodule, attr)
-
-__all__ = getattr(_submodule, '__all__', [])
+__all__ = ["HCSEMixin", "HfTrainerWithHCSE"]
