+++ holland_dual/cognition/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from .hcse.core import HCSEMixin
-from .hcse.pipeline import HfTrainerWithHCSE
+from hcse.core import HCSEMixin
+from hcse.pipeline import HfTrainerWithHCSE
 
 __all__ = ["HCSEMixin", "HfTrainerWithHCSE"]
