--- holland_dual/cognition/hcse/__init__.py
@@ -1,6 +1,0 @@
-"""Holland Consciousness Scaling Engine package."""
-
-from .core import HCSEMixin
-from .pipeline import HfTrainerWithHCSE
-
-__all__ = ["HCSEMixin", "HfTrainerWithHCSE"]
